Re-add word_count resetting on newline
This is necessary to reset alias checking, since aliases can only be
expanded if they are the first word in a command. It needs to be handled
in the lexer because the result has to be relexed, which can only happen
if we are keeping track of the word count in the lexer.